@@ -1,65 +1,42 @@
 {
   "name": "mussel",
-  "version": "1.0.5",
-  "description": "Another UI Library for Vue",
-  "main": "dist/mussel.esm.js",
+  "version": "4.0.0-alpha.5",
+  "description": "UI Library for Vue@3",
+  "main": "dist/mussel.js",
+  "type": "module",
   "files": [
-    "dist",
-    "src"
+    "dist"
   ],
   "scripts": {
-    "build": "rollup -c",
-    "build:dev": "rollup -c --watch --environment=dev",
-    "docs": "concurrently \"node open-docs.js\" \"serve docs\"",
+    "build": "rollup -c --environment=dev && rollup -c",
+    "build:dev": "rollup -c --environment=dev --watch",
     "pub": "npm run build && npm publish"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/eslint-parser": "^7.14.3",
-    "@babel/preset-env": "^7.14.2",
-    "@rollup/plugin-alias": "^3.1.2",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^19.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@tabler/icons": "^1.41.2",
-    "@utilitycss/postcss-for": "^0.0.4",
-    "@vue/babel-plugin-transform-vue-jsx": "^1.2.1",
-    "a-color-picker": "^1.2.1",
-    "autoprefixer": "^10.3.1",
-    "concurrently": "^5.3.0",
-    "core-js": "^3.12.1",
-    "eslint": "^7.26.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.23.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-vue": "^7.9.0",
-    "lodash.debounce": "^4.0.8",
-    "lodash.isdate": "^4.0.1",
-    "lodash.isfunction": "^3.0.9",
-    "lodash.isnil": "^4.0.0",
-    "lodash.isplainobject": "^4.0.6",
-    "lodash.isstring": "^4.0.1",
-    "lodash.throttle": "^4.1.1",
-    "open": "^8.0.9",
-    "param-case": "^3.0.4",
-    "postcss": "^8.3.5",
-    "postcss-advanced-variables": "^3.0.1",
-    "postcss-calc": "^8.0.0",
-    "postcss-nested": "^5.0.5",
-    "resize-observer-polyfill": "^1.5.1",
-    "rollup": "^2.48.0",
-    "rollup-plugin-postcss": "^4.0.0",
+    "@cheap-pets/rollup-plugin-postcss-scss": "^1.0.2",
+    "@rollup/plugin-alias": "^5.1.1",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@swc/core": "^1.7.26",
+    "@tabler/icons": "^3.17.0",
+    "@tabler/icons-webfont": "^3.17.0",
+    "browserslist": "^4.23.3",
+    "core-js": "^3.38.1",
+    "mussel-scrollbar": "^1.0.5",
+    "postcss-simple-vars": "^7.0.1",
+    "rollup": "^4.22.4",
     "rollup-plugin-string": "^3.0.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-vue": "^5.1.9",
-    "serve": "^11.3.2",
-    "vue": "^2.6.12",
-    "vue-template-compiler": "^2.6.12"
+    "rollup-plugin-swc": "^0.2.1",
+    "throttle-debounce": "^5.0.2",
+    "unplugin-vue": "^5.1.5",
+    "vite": "^5.4.7",
+    "vue": "^3.5.8"
   },
-  "repository": "https://github.com/cheap-pets/mussel.git",
+  "config": {
+    "cz-customizable": {
+      "config": ".cz-config.cjs"
+    }
+  },
+  "repository": "https://github.com/cheap-pets/mussel3.git",
   "author": "scrollbar-ww <scrollbar@163.com>",
   "license": "MIT"
 }
