@@ -1,49 +1,18 @@
 <template>
-  <div
-    class="mu-flex-box mu-form"
-    :form-style="formStyle"
-    :cellpadding="cellpadding">
+  <div class="mu-form" :style="style">
     <slot />
   </div>
 </template>
 
-<script>
-  import FlexBox from '../layout/flex-box.vue'
+<script setup>
+  import { provide } from 'vue'
+  import { sizeProps, useSize } from '@/hooks/size'
 
-  export default {
-    name: 'MusselForm',
-    extends: FlexBox,
-    provide () {
-      return {
-        form: this
-      }
-    },
-    props: {
-      layout: {
-        type: String,
-        default: 'flow',
-        validator (value) {
-          return ['flow', 'column', 'row'].indexOf(value) !== -1
-        }
-      },
-      formStyle: {
-        type: String,
-        default: 'normal',
-        validator (value) {
-          return ['normal', 'table'].indexOf(value) !== -1
-        }
-      },
-      cellpadding: {
-        type: Boolean,
-        default () {
-          return this.formStyle !== 'table'
-        }
-      },
-      labelWidth: {
-        type: String,
-        default: '75px'
-      },
-      labelAlign: String
-    }
-  }
+  defineOptions({ name: 'MusselForm' })
+
+  const props = defineProps({ ...sizeProps, labelWidth: String, labelAlign: String })
+  const style = useSize(props).sizeStyle
+
+  provide('form', props)
+
 </script>
