@@ -1,5 +1,12 @@
-import './message.pcss'
+import { pluginNotifier } from './notifier.js'
+import { pluginMessageBox } from './message-box.js'
 
-export { showMessage, alert, confirm, error, warn } from './message-box'
-export { notify } from './notify'
-export * from './prompt-panel.vue'
+function install (app, options = {}) {
+  app.config.globalProperties.$mussel.messageBox = {
+    ...pluginNotifier(app),
+    ...pluginMessageBox(app)
+  }
+}
+
+export { default as MuStatusBox } from './status-box.vue'
+export { install }
