@@ -1,34 +1,31 @@
-import ResizeObserver from 'resize-observer-polyfill'
+import { EventInterceptor } from './interceptor'
+import { dispatchCustomEvent } from './custom-event'
 
-import emit from './emit'
-import interceptor from './interceptor'
+const OBSERVE_COUNT_SYMBOL = Symbol.for('mussel.event.resize.count')
 
-const GLOBAL_OBSERVER_NAME = '__mussel_resize_observer'
-const LISTENER_COUNT_PROP_NAME = '__mussel_resize_listener_count'
-
-function getResizeObserver () {
-  if (!window[GLOBAL_OBSERVER_NAME]) {
-    window[GLOBAL_OBSERVER_NAME] = new ResizeObserver(entries => {
-      entries.forEach(entry => emit(entry.target, 'sizechange'))
-    })
-  }
-  return window[GLOBAL_OBSERVER_NAME]
-}
-
-function interceptorAdd () {
-  const count = this[LISTENER_COUNT_PROP_NAME] || 0
-  if (!count) getResizeObserver().observe(this)
-
-  this[LISTENER_COUNT_PROP_NAME] = count + 1
-}
-
-function interceptorRemove () {
-  const count = this[LISTENER_COUNT_PROP_NAME] || 0
-  this[LISTENER_COUNT_PROP_NAME] = count ? count - 1 : 0
-  if (!count) getResizeObserver().unobserve(this)
-}
-
-interceptor.register('sizechange', {
-  add: interceptorAdd,
-  remove: interceptorRemove
+const observer = new ResizeObserver(entries => {
+  entries.forEach(entry => dispatchCustomEvent(entry.target, 'sizechange'))
 })
+
+const interceptor = {
+  add () {
+    const count = this[OBSERVE_COUNT_SYMBOL] || 0
+
+    if (!count) observer.observe(this)
+
+    this[OBSERVE_COUNT_SYMBOL] = count + 1
+  },
+
+  remove () {
+    const count = this[OBSERVE_COUNT_SYMBOL] || 0
+
+    if (count <= 1) {
+      this[OBSERVE_COUNT_SYMBOL] = 0
+      observer.unobserve(this)
+    } else {
+      this[OBSERVE_COUNT_SYMBOL]--
+    }
+  }
+}
+
+EventInterceptor.register('sizechange', interceptor)
